pyproject.toml
[project]
name = "geometrikks"
version = "0.1.0"
description = "Geo metrics ingress service built with Litestar."
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "litestar[standard,brotli,jwt,picologging,prometheus,pydantic,redis,sqlalchemy]>=2.18.0",
    "geoip2>=5.2.0",
    "geohash2>=1.1",
    "ipy>=1.1",
    "python-dotenv>=1.2.1",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.30.0",
    "httpx-oauth>=0.16.1",
    "aiofiles>=25.1.0",
    "geoalchemy2>=0.18.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    # Reduce noise from third-party libs; keep real test warnings visible
    "ignore:There is no current event loop:DeprecationWarning",
]
